added chat model and some insertions

+++ app/models/equipment_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, Integer, String, DateTime, LargeBinary, JSON, ForeignKey
+from sqlalchemy import Column, BigInteger, String, DateTime, LargeBinary, JSON, ForeignKey
 from datetime import datetime
 from app.config.db import Base
 
@@ -7,12 +7,12 @@ from app.config.db import Base
 # Stores uploaded equipment images and their embeddings
 # -----------------------------
 class EquipmentImage(Base):
-    __tablename__ = "equipment_embeddings"  # Table name in DB
+    __tablename__ = "equipment_embeddings"
 
     id = Column(BigInteger, primary_key=True, autoincrement=True)
-    filename = Column(String(255), nullable=False) 
+    filename = Column(String(255), nullable=False)
     uploaded_at = Column(DateTime, default=datetime.utcnow)
-    vector = Column(LargeBinary, nullable=False)       
+    vector = Column(LargeBinary, nullable=False)
     extra_metadata = Column("metadata", JSON)
 
-    chat_session_id = Column(Integer, ForeignKey("chat_sessions.id"))
+    chat_session_id = Column(BigInteger, ForeignKey("chat_sessions.id"))
